showroom/usecase-011/requirements.txt
PyJWT>=2.8.0
cryptography>=44.0.0
requests>=2.32.0
python-dotenv>=1.0.0
flask>=2.3.0
pyngrok>=7.0.0
openai>=1.12.0
python-dateutil>=2.8.2
tiktoken>=0.5.2
langchain>=0.1.2
langchain-community>=0.0.13
langchain-openai>=0.0.5
faiss-cpu>=1.7.4
pypdf>=4.0.1
python-docx>=1.0.1
markdown>=3.5
pydantic>=2.5.2
tenacity>=8.2.3